Use new DAK headers to identify "Accepted" mails
This should avoid some issues with folding of the subject header where
the "source" string was not on the same line as the "Accepted" line.

Note that we still need to parse the subject to differentiate between
"Accepted $pkg $vers (source) into unstable" and "$changes ACCEPTED into
unstable". Only the former generates a news entry.