test: properly reset cached values when overriding settings
The FileSystemStorage object caches the value of MEDIA_ROOT
but it properly connects to the setting_changed signal to invalidate
cached values so we just have to emit that signal.

This is the reason why extract_source_files tests are working
individually but are failing when run together since the actual
path used corresponds to the MEDIA_ROOT value for the first test
while the code using settings.MEDIA_ROOT has another path.