Restart migrations from scratch to get rid of jsonfield dependency
We add the "replaces" to indicate that we are squashing together the
former migrations. However we immediately remove the old migrations
because they are the source of the dependency that we want to get rid
of.

So instead of the progressive replacement that is usually expected
when you squash migrations together, we have to force a full upgrade
to the former version and we document this in the upgrade notes.